requirements.txt
asgiref==3.2.7
at==0.0.3
autopep8==1.5.1
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
DateTime==4.3
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-unused-media==0.1.13
fake-useragent==0.1.11
gunicorn==20.0.4
idna==2.9
lxml==4.5.0
mysql==0.0.2
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.1.1
pkg-resources==0.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
python3-openid==3.1.0
pytz==2019.3
requests==2.10.0
requests-oauthlib==1.3.0
rutimeparser==1.1.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
telebot==0.0.3
urllib3==1.25.8
zope.interface==5.0.1
